requirements.txt
bing_image_downloader==1.1.2
matplotlib==3.7.5
numpy==1.24.3
pillow==10.3.0
scikit-learn==1.3.2
tensorflow_probability==0.19.0
# (tensorflow itself installed separately for cpu or gpu version)
